promote shared types to common
