+++ packages/sdk/src/index.ts
@@ -82,7 +82,6 @@ export type {
   LendMarketMetadata,
   LendMarketPosition,
   LendMarketSupply,
-  LendOptions,
   LendProviderConfig,
   LendTransaction,
   LendTransactionReceipt,
@@ -103,6 +102,7 @@ export type {
   SwapTransaction,
   TokenBalance,
   TransactionData,
+  TransactionOptions,
   UserOperationTransactionReceipt,
   WalletConfig,
   WalletSwapParams,
