+++ packages/sdk/src/types/index.ts
@@ -1,6 +1,7 @@
 export * from '@/types/actions.js'
 export * from '@/types/asset.js'
 export * from '@/types/chain.js'
+export * from '@/types/common/index.js'
 export * from '@/types/lend/index.js'
 export * from '@/types/swap/index.js'
 export * from '@/types/transaction.js'
